Create 1_Machine_Learning_Project_Honey_Production.py
Now that you have learned how linear regression works, let’s try it on an example of real-world data.
As you may have already heard, the honeybees are in a precarious state right now. You may have seen articles about the decline of the honeybee population for various reasons. You want to investigate this decline and how the trends of the past predict the future for the honeybees.
Note: All the tasks can be completed using Pandas or NumPy. Pick whichever one you prefer.
We have loaded in a DataFrame for you about honey production in the United States from Kaggle. It is called df.
For now, we care about the total production of honey per year.
Create a variable called X that is the column of years in this prod_per_year DataFrame.
Plot y vs X as a scatterplot.
Create a linear regression model from scikit-learn and call it regr.
Fit the model to the data.
After you have fit the model, print out the slope of the line and the intercept of the line.
Create a list called y_predict that is the predictions your regr model would make on the X data.
Plot y_predict vs X as a line, on top of your scatterplot.
So, it looks like the production of honey has been in decline, according to this linear model. Let’s predict what the year 2050 may look like in terms of honey production.
Our known dataset stops at the year 2013, so let’s create a NumPy array called X_future that is the range from 2013 to 2050. 
After creating that array, we need to reshape it for scikit-learn.
Plot future_predict vs X_future.
How much honey will be produced in the year 2050, according to this?

Ans: Around 100000