Update 3_Predict_Titanic_Survival_logistic_regression.py
Predict Titanic Survival
The RMS Titanic set sail on its maiden voyage in 1912, crossing the Atlantic from Southampton, England to New York City. The ship never completed the voyage, sinking to the bottom of the Atlantic Ocean after hitting an iceberg, bringing down 1,502 of 2,224 passengers onboard.
In this project I created a Logistic Regression model that predicts which passengers survived the sinking of the Titanic, based on features like age and class.